+++ core/property.py
